Make Neuron Scraper extension for simple prometheus scraper
