+++ receiver/awscontainerinsightreceiver/config.go
@@ -61,7 +61,6 @@ type Config struct {
 	// EnableGpuMetric toggles GPU monitoring where metrics are scraped from vendor specific sources
 	EnableGpuMetric bool `mapstructure:"gpu_metrics"`
 
-	// EnableNeuronMetric disables Neuron monitoring where metrics are scraped from vendor specific sources
-	// The default value is true meaning Neuron metrics get collected out of the box unless it's disabled
+	// EnableNeuronMetric toggles Neuron monitoring where metrics are scraped from vendor specific sources
 	EnableNeuronMetric bool `mapstructure:"neuron_metrics"`
 }
