+++ receiver/awscontainerinsightreceiver/internal/neuron/neuron_monitor_scraper_test.go
@@ -1,24 +1,11 @@
-package neuron
+package nueron
 
 import (
-	"context"
-	"fmt"
 	"strings"
 	"testing"
 
-	"github.com/open-telemetry/opentelemetry-collector-contrib/receiver/awscontainerinsightreceiver/internal/mocks"
-	"github.com/open-telemetry/opentelemetry-collector-contrib/receiver/prometheusreceiver"
-	configutil "github.com/prometheus/common/config"
-	"github.com/prometheus/common/model"
-	"github.com/prometheus/prometheus/config"
-	"github.com/prometheus/prometheus/discovery"
-	"github.com/prometheus/prometheus/model/relabel"
+	"github.com/open-telemetry/opentelemetry-collector-contrib/receiver/awscontainerinsightreceiver/internal/prometheusscraper"
 	"github.com/stretchr/testify/assert"
-	"go.opentelemetry.io/collector/component/componenttest"
-	"go.opentelemetry.io/collector/consumer"
-	"go.opentelemetry.io/collector/pdata/pmetric"
-	"go.opentelemetry.io/collector/receiver"
-	"go.uber.org/zap"
 )
 
 const renameMetric = `
@@ -105,289 +92,69 @@ system_vcpu_usage_ratio{availability_zone="us-east-1c",instance_id="i-09db9b55e0
 instance_info{availability_zone="us-east-1c",instance_id="i-09db9b55e0095612f",instance_name="",instance_type="trn1n.32xlarge",region="us-east-1",subnet_id="subnet-06a7754948e8a000f"} 1.0
 `
 
-const dummyInstanceId = "i-0000000000"
-const nueronCoreID = "0"
+const dummyClusterName = "cluster-name"
+const dummyHostName = "i-000000000"
 
 type mockHostInfoProvider struct {
 }
 
 func (m mockHostInfoProvider) GetClusterName() string {
-	return "cluster-name"
+	return dummyClusterName
 }
 
 func (m mockHostInfoProvider) GetInstanceID() string {
-	return dummyInstanceId
-}
-
-type mockConsumer struct {
-	t             *testing.T
-	up            *bool
-	coreUtil      *bool
-	memUsed       *bool
-	instanceId    *bool
-	podName       *bool
-	containerName *bool
-	namespace     *bool
-	relabeled     *bool
-}
-
-func (m mockConsumer) Capabilities() consumer.Capabilities {
-	return consumer.Capabilities{
-		MutatesData: false,
-	}
-}
-
-func (m mockConsumer) ConsumeMetrics(_ context.Context, md pmetric.Metrics) error {
-	assert.Equal(m.t, 1, md.ResourceMetrics().Len())
-
-	scopeMetrics := md.ResourceMetrics().At(0).ScopeMetrics().At(0).Metrics()
-	for i := 0; i < scopeMetrics.Len(); i++ {
-		metric := scopeMetrics.At(i)
-		if metric.Name() == "neuroncore_utilization_ratio" {
-			assert.Equal(m.t, float64(0.1), metric.Gauge().DataPoints().At(0).DoubleValue())
-			*m.coreUtil = true
-		}
-
-		if strings.Contains(metric.Name(), "neuroncore") {
-			instanceId, _ := metric.Gauge().DataPoints().At(0).Attributes().Get("InstanceId")
-			podName, _ := metric.Gauge().DataPoints().At(0).Attributes().Get("PodName")
-			namespace, _ := metric.Gauge().DataPoints().At(0).Attributes().Get("Namespace")
-			containerName, _ := metric.Gauge().DataPoints().At(0).Attributes().Get("Namespace")
-			_, relabeled := metric.Gauge().DataPoints().At(0).Attributes().Get("DeviceId")
-			*m.instanceId = instanceId.Str() == "i-09db9b55e0095612f"
-			*m.podName = podName.Str() == nueronCoreID
-			*m.namespace = namespace.Str() == nueronCoreID
-			*m.containerName = containerName.Str() == nueronCoreID
-			*m.relabeled = relabeled
-		}
-
-		if strings.Contains(metric.Name(), "system") {
-			instanceId, _ := metric.Gauge().DataPoints().At(0).Attributes().Get("InstanceId")
-			_, podNameFound := metric.Gauge().DataPoints().At(0).Attributes().Get("PodName")
-			_, namespaceFound := metric.Gauge().DataPoints().At(0).Attributes().Get("Namespace")
-			_, relabelFound := metric.Gauge().DataPoints().At(0).Attributes().Get("DeviceId")
-			assert.Equal(m.t, instanceId.Str(), "i-09db9b55e0095612f")
-			assert.False(m.t, podNameFound)
-			assert.False(m.t, namespaceFound)
-			assert.False(m.t, relabelFound)
-		}
-
-		if metric.Name() == "neuron_runtime_memory_used_bytes" {
-			assert.Equal(m.t, float64(9.043968e+06), metric.Gauge().DataPoints().At(0).DoubleValue())
-			*m.memUsed = true
-		}
-
-		if metric.Name() == "up" {
-			assert.Equal(m.t, float64(1.0), metric.Gauge().DataPoints().At(0).DoubleValue())
-			*m.up = true
-		}
-	}
-
-	return nil
-}
-
-func TestNewNeuronScraperBadInputs(t *testing.T) {
-	settings := componenttest.NewNopTelemetrySettings()
-	settings.Logger, _ = zap.NewDevelopment()
-
-	tests := []NeuronMonitorScraperOpts{
-		{
-			Ctx:               context.TODO(),
-			TelemetrySettings: settings,
-			Consumer:          nil,
-			Host:              componenttest.NewNopHost(),
-			HostInfoProvider:  mockHostInfoProvider{},
-		},
-		{
-			Ctx:               context.TODO(),
-			TelemetrySettings: settings,
-			Consumer:          mockConsumer{},
-			Host:              nil,
-			HostInfoProvider:  mockHostInfoProvider{},
-		},
-		{
-			Ctx:               context.TODO(),
-			TelemetrySettings: settings,
-			Consumer:          mockConsumer{},
-			Host:              componenttest.NewNopHost(),
-			HostInfoProvider:  nil,
-		},
-	}
-
-	for _, tt := range tests {
-		scraper, err := NewNeuronMonitorScraper(tt)
-
-		assert.Error(t, err)
-		assert.Nil(t, scraper)
-	}
+	return dummyHostName
 }
 
 func TestNewNeuronScraperEndToEnd(t *testing.T) {
-
-	upPtr := false
-	coreUtil := false
-	instanceId := false
-	memUsed := false
-	podName := false
-	containerName := false
-	namespace := false
-	relabeled := false
-
-	consumer := mockConsumer{
-		t:             t,
-		up:            &upPtr,
-		coreUtil:      &coreUtil,
-		memUsed:       &memUsed,
-		instanceId:    &instanceId,
-		podName:       &podName,
-		containerName: &containerName,
-		namespace:     &namespace,
-		relabeled:     &relabeled,
+	expectedMetrics := make(map[string]prometheusscraper.ExpectedMetricStruct)
+	expectedMetrics["neuroncore_utilization_ratio"] = prometheusscraper.ExpectedMetricStruct{
+		MetricValue: 0.1,
+		MetricLabels: []prometheusscraper.MetricLabel{
+			{LabelName: "InstanceId", LabelValue: "i-09db9b55e0095612f"},
+			{LabelName: "ClusterName", LabelValue: dummyClusterName},
+			{LabelName: "DeviceId", LabelValue: "0"},
+		},
+	}
+	expectedMetrics["neuron_runtime_memory_used_bytes"] = prometheusscraper.ExpectedMetricStruct{
+		MetricValue: 9.043968e+06,
+		MetricLabels: []prometheusscraper.MetricLabel{
+			{LabelName: "InstanceId", LabelValue: "i-09db9b55e0095612f"},
+			{LabelName: "ClusterName", LabelValue: dummyClusterName},
+		},
 	}
 
-	settings := componenttest.NewNopTelemetrySettings()
-	settings.Logger, _ = zap.NewDevelopment()
+	expectedMetrics["execution_errors_created"] = prometheusscraper.ExpectedMetricStruct{
+		MetricValue: 1.7083389404380567e+09,
+		MetricLabels: []prometheusscraper.MetricLabel{
+			{LabelName: "InstanceId", LabelValue: "i-09db9b55e0095612f"},
+			{LabelName: "ClusterName", LabelValue: dummyClusterName},
+		},
+	}
 
-	scraper, err := NewNeuronMonitorScraper(NeuronMonitorScraperOpts{
-		Ctx:               context.TODO(),
-		TelemetrySettings: settings,
-		Consumer:          mockConsumer{},
-		Host:              componenttest.NewNopHost(),
-		HostInfoProvider:  mockHostInfoProvider{},
+	expectedMetrics["up"] = prometheusscraper.ExpectedMetricStruct{
+		MetricValue:  1.0,
+		MetricLabels: []prometheusscraper.MetricLabel{},
+	}
+
+	consumer := prometheusscraper.MockConsumer{
+		T:               t,
+		ExpectedMetrics: expectedMetrics,
+	}
+
+	mockedScraperOpts := prometheusscraper.GetMockedScraperOpts(consumer, mockHostInfoProvider{})
+
+	prometheusscraper.TestSimplePrometheusEndToEnd(prometheusscraper.TestSimplePrometheusEndToEndOpts{
+		T:                   t,
+		Consumer:            consumer,
+		DataReturned:        renameMetric,
+		ScraperOpts:         mockedScraperOpts,
+		ScrapeConfig:        GetNueronScrapeConfig(mockedScraperOpts),
+		MetricRelabelConfig: GetNueronMetricRelabelConfigs(mockedScraperOpts),
 	})
-	assert.NoError(t, err)
-	assert.Equal(t, mockHostInfoProvider{}, scraper.hostInfoProvider)
-
-	// build up a new PR
-	promFactory := prometheusreceiver.NewFactory()
-
-	targets := []*mocks.TestData{
-		{
-			Name: "neuron",
-			Pages: []mocks.MockPrometheusResponse{
-				{Code: 200, Data: renameMetric},
-			},
-		},
-	}
-	mp, cfg, err := mocks.SetupMockPrometheus(targets...)
-	assert.NoError(t, err)
-
-	split := strings.Split(mp.Srv.URL, "http://")
-
-	scrapeConfig := &config.ScrapeConfig{
-		HTTPClientConfig: configutil.HTTPClientConfig{
-			TLSConfig: configutil.TLSConfig{
-				InsecureSkipVerify: true,
-			},
-		},
-		ScrapeInterval:  cfg.ScrapeConfigs[0].ScrapeInterval,
-		ScrapeTimeout:   cfg.ScrapeConfigs[0].ScrapeInterval,
-		JobName:         fmt.Sprintf("%s/%s", jobName, cfg.ScrapeConfigs[0].MetricsPath),
-		HonorTimestamps: true,
-		Scheme:          "http",
-		MetricsPath:     cfg.ScrapeConfigs[0].MetricsPath,
-		ServiceDiscoveryConfigs: discovery.Configs{
-			// using dummy static config to avoid service discovery initialization
-			&discovery.StaticConfig{
-				{
-					Targets: []model.LabelSet{
-						{
-							model.AddressLabel: model.LabelValue(split[1]),
-						},
-					},
-				},
-			},
-		},
-		RelabelConfigs: []*relabel.Config{},
-		MetricRelabelConfigs: []*relabel.Config{
-			{
-				SourceLabels: model.LabelNames{"__name__"},
-				Regex:        relabel.MustNewRegexp("neuron.*|system_.*|execution_.*"),
-				Action:       relabel.Keep,
-			},
-			{
-				SourceLabels: model.LabelNames{"instance_id"},
-				TargetLabel:  "InstanceId",
-				Regex:        relabel.MustNewRegexp("(.*)"),
-				Replacement:  "${1}",
-				Action:       relabel.Replace,
-			},
-			{
-				SourceLabels: model.LabelNames{"neuroncore"},
-				TargetLabel:  "DeviceId",
-				Regex:        relabel.MustNewRegexp("(.*)"),
-				Replacement:  "${1}",
-				Action:       relabel.Replace,
-			},
-			// test hack to inject cluster name as label
-			{
-				SourceLabels: model.LabelNames{"instance_id"},
-				TargetLabel:  "ClusterName",
-				Regex:        relabel.MustNewRegexp("neuron.*"),
-				Replacement:  scraper.hostInfoProvider.GetClusterName(),
-				Action:       relabel.Replace,
-			},
-			{
-				SourceLabels: model.LabelNames{"neuroncore"},
-				TargetLabel:  "Namespace",
-				Regex:        relabel.MustNewRegexp("(.*)"),
-				Replacement:  "${1}",
-				Action:       relabel.Replace,
-			},
-			{
-				SourceLabels: model.LabelNames{"neuroncore"},
-				TargetLabel:  "PodName",
-				Regex:        relabel.MustNewRegexp("(.*)"),
-				Replacement:  "${1}",
-				Action:       relabel.Replace,
-			},
-			{
-				SourceLabels: model.LabelNames{"neuroncore"},
-				TargetLabel:  "ContainerName",
-				Regex:        relabel.MustNewRegexp("(.*)"),
-				Replacement:  "${1}",
-				Action:       relabel.Replace,
-			},
-		},
-	}
-
-	promConfig := prometheusreceiver.Config{
-		PrometheusConfig: &config.Config{
-			ScrapeConfigs: []*config.ScrapeConfig{scrapeConfig},
-		},
-	}
-
-	// replace the prom receiver
-	params := receiver.CreateSettings{
-		TelemetrySettings: scraper.settings,
-	}
-	scraper.prometheusReceiver, err = promFactory.CreateMetricsReceiver(scraper.ctx, params, &promConfig, consumer)
-	assert.NoError(t, err)
-	assert.NotNil(t, mp)
-	defer mp.Close()
-
-	// perform a single scrape, this will kick off the scraper process for additional scrapes
-	scraper.GetMetrics()
-
-	t.Cleanup(func() {
-		scraper.Shutdown()
-	})
-
-	// wait for 2 scrapes, one initiated by us, another by the new scraper process
-	mp.Wg.Wait()
-	mp.Wg.Wait()
-
-	assert.True(t, *consumer.up)
-	assert.True(t, *consumer.coreUtil)
-	assert.True(t, *consumer.memUsed)
-	assert.True(t, *consumer.instanceId)
-	assert.True(t, *consumer.podName)
-	assert.True(t, *consumer.containerName)
-	assert.True(t, *consumer.namespace)
-	assert.True(t, *consumer.relabeled)
 }
 
-func TestDcgmScraperJobName(t *testing.T) {
+func TestNeuronMonitorScraperJobName(t *testing.T) {
 	// needs to start with containerInsights
 	assert.True(t, strings.HasPrefix(jobName, "containerInsightsNeuronMonitorScraper"))
 }
